Moved SnapshotVersion component to its own header
